+++ src/MemoryPack.Generator/MemoryPackGenerator.Emitter.cs
@@ -852,7 +852,7 @@ partial {{classOrInterface}} {{TypeName}} : IMemoryPackFormatterRegister
         var elements = UnionTags.Select(x => $"            {{ typeof({ToUnionTagTypeFullyQualifiedToString(x.Type)}), {x.Tag} }},").NewLine();
 
         return $$"""
-        static readonly System.Collections.Generic.Dictionary<Type, byte> __typeToTag = new({{UnionTags.Length}})
+        static readonly System.Collections.Generic.Dictionary<Type, ushort> __typeToTag = new({{UnionTags.Length}})
         {
 {{elements}}
         };
